+++ tests/Bridge/MoodlePluginCollectionTest.php
@@ -49,4 +49,34 @@ class MoodlePluginCollectionTest extends \PHPUnit_Framework_TestCase
         $this->assertSame($plugin2, $all[1]);
         $this->assertSame($plugin1, $all[2]);
     }
+
+    /**
+     * Moodle plugins can have circular dependencies, ensure we do not error when this happens.
+     */
+    public function testSortByDependenciesCircularError()
+    {
+        $plugin1               = new DummyMoodlePlugin('');
+        $plugin1->component    = 'mod_1';
+        $plugin1->dependencies = ['mod_2', 'mod_3'];
+
+        $plugin2               = new DummyMoodlePlugin('');
+        $plugin2->component    = 'mod_2';
+        $plugin2->dependencies = ['mod_1'];
+
+        $plugin3            = new DummyMoodlePlugin('');
+        $plugin3->component = 'mod_3';
+
+        $plugins = new MoodlePluginCollection();
+        $plugins->add($plugin1);
+        $plugins->add($plugin3);
+        $plugins->add($plugin2);
+
+        $sorted = $plugins->sortByDependencies();
+        $all    = $sorted->all();
+
+        $this->assertCount(3, $all);
+        $this->assertSame($plugin2, $all[0]);
+        $this->assertSame($plugin3, $all[1]);
+        $this->assertSame($plugin1, $all[2]);
+    }
 }